experiments/simclr-experiments/requirements.txt
lightly
tensorboard
numpy
scikit-image
matplotlib
torch-summary